# Fix escaped strings handling