--- a/src/field.jl
+++ b/src/field.jl
@@ -295,6 +295,7 @@ show(io::IO, c::StringToken) = print(io, "<string>")
 fromtype(::Type{S}) where {S<:AbstractString} = StringToken(S)
 
 function tryparsenext(s::StringToken{T}, str, i, len, opts) where {T}
+    escapecount = 0
     R = Nullable{T}
     p = ' '
     i0 = i
@@ -312,6 +313,11 @@ function tryparsenext(s::StringToken{T}, str, i, len, opts) where {T}
     y2 = iterate(str, i)
     while y2!==nothing
         c = y2[1]; ii = y2[2]
+
+        if p==opts.escapechar
+            escapecount += 1
+        end
+
         if opts.spacedelim && (c == ' ' || c == '\t')
             break
         elseif !opts.spacedelim && c == Char(opts.endchar)
@@ -358,14 +364,28 @@ function tryparsenext(s::StringToken{T}, str, i, len, opts) where {T}
         y2 = iterate(str, i)
     end
 
-    return R(_substring(T, str, i0, i-1)), i
+    return R(_substring(T, str, i0, i-1, escapecount, opts.escapechar)), i
 end
 
-@inline function _substring(::Type{String}, str, i, j)
-    String(str[i:thisind(str, j)])
+@inline function _substring(::Type{String}, str, i, j, escapecount, escapechar)
+    if escapecount > 0
+        buf = IOBuffer()
+        cur_i = i
+        while cur_i < j
+            c = str[cur_i] 
+            if c != escapechar
+                print(buf, c)
+            end
+            cur_i = nextind(str, cur_i)
+        end
+        return String(take!(buf))
+    else
+        return String(str[i:thisind(str, j)])
+    end
 end
 
-@inline function _substring(::Type{T}, str, i, j) where {T<:SubString}
+@inline function _substring(::Type{T}, str, i, j, escapecount, escapechar) where {T<:SubString}
+    escapecount > 0 && error("Not yet handled 2")
     T(str, i, thisind(j))
 end
 
@@ -375,11 +395,12 @@ fromtype(::Type{StrRange}) = StringToken(StrRange)
     unsafe_string(pointer(str, 1 + r.offset), r.length)
 end
 
-@inline function _substring(::Type{StrRange}, str, i, j)
-    StrRange(i - 1, j - i + 1)
+@inline function _substring(::Type{StrRange}, str, i, j, escapecount, escapechar)
+    StrRange(i - 1, j - i + 1, escapecount)
 end
 
-@inline function _substring(::Type{<:WeakRefString}, str, i, j)
+@inline function _substring(::Type{<:WeakRefString}, str, i, j, escapecount, escapechar)
+    escapecount > 0 && error("Not yet handled 3")
     WeakRefString(convert(Ptr{UInt8}, pointer(str, i)), j - i + 1)
 end