--- a/src/field.jl
+++ b/src/field.jl
@@ -364,16 +364,24 @@ function tryparsenext(s::StringToken{T}, str, i, len, opts) where {T}
         y2 = iterate(str, i)
     end
 
-    return R(_substring(T, str, i0, i-1, escapecount, opts.escapechar)), i
+    return R(_substring(T, str, i0, i-1, escapecount, opts.escapechar, opts.quotechar)), i
 end
 
-@inline function _substring(::Type{String}, str, i, j, escapecount, escapechar)
+@inline function _substring(::Type{String}, str, i, j, escapecount, escapechar, quotechar)
     if escapecount > 0
         buf = IOBuffer()
         cur_i = i
-        while cur_i < j
+        while cur_i <= j
             c = str[cur_i] 
-            if c != escapechar
+            if c == escapechar
+                next_i = nextind(str, cur_i)
+                if next_i <= j && str[next_i] == quotechar
+                    print(buf, str[next_i])
+                    cur_i = next_i
+                else
+                    print(buf, c)
+                end
+            else
                 print(buf, c)
             end
             cur_i = nextind(str, cur_i)
@@ -384,7 +392,7 @@ end
     end
 end
 
-@inline function _substring(::Type{T}, str, i, j, escapecount, escapechar) where {T<:SubString}
+@inline function _substring(::Type{T}, str, i, j, escapecount, escapechar, quotechar) where {T<:SubString}
     escapecount > 0 && error("Not yet handled 2")
     T(str, i, thisind(j))
 end
@@ -395,11 +403,11 @@ fromtype(::Type{StrRange}) = StringToken(StrRange)
     unsafe_string(pointer(str, 1 + r.offset), r.length)
 end
 
-@inline function _substring(::Type{StrRange}, str, i, j, escapecount, escapechar)
+@inline function _substring(::Type{StrRange}, str, i, j, escapecount, escapechar, quotechar)
     StrRange(i - 1, j - i + 1, escapecount)
 end
 
-@inline function _substring(::Type{<:WeakRefString}, str, i, j, escapecount, escapechar)
+@inline function _substring(::Type{<:WeakRefString}, str, i, j, escapecount, escapechar, quotechar)
     escapecount > 0 && error("Not yet handled 3")
     WeakRefString(convert(Ptr{UInt8}, pointer(str, i)), j - i + 1)
 end