--- a/src/record.jl
+++ b/src/record.jl
@@ -96,7 +96,7 @@ end
 end
 
 @inline Base.@propagate_inbounds function setcell!(col::StringVector, i, val::StrRange, str)
-    val.escapecount>0 && @debug("Not yet implemented")
+    # TODO Properly handle the val.escapecount>0 case
     col[i] = WeakRefString(pointer(str, val.offset + 1), val.length)
     PARSE_SUCCESS
 end