--- a/src/util.jl
+++ b/src/util.jl
@@ -316,6 +316,7 @@ failedat(xs) = (@assert isnull(xs[1]); xs[2])
 struct StrRange
     offset::Int
     length::Int
+    escapecount::Int
 end
 
 function getlineat(str, i)