Fixes #85.

@shashi could you take a look whether you are ok with the general strategy here? It is not yet done, but it would be good to get some feedback before I continue with this one here.

I generally have changed how the case of `Vector{String}` is handled: in that case this PR is no longer using `StrRange` at all, but is allocating the `String` in the `tryparsenext` method itself. I didn't see a benefit of going through `StrRange`, and it cuts down on the places that need to handle the escaping stuff.

This does *not yet* work with `StringArray`. I'm not sure I'll have the time to implement this for `StringArray` as well, in particular given that I'm not using it at all. Maybe someone from the JuliaDB team could sort out that part? Or, alternatively, but clearly not ideal, we could also for now just fix the bug for the `Vector{String}` case and then later try to fix it for `StringArray` as well.

Oh, and it would be great if you could carefully look over my changes to the tests. I think they now reflect a world in which escape char is handled properly, but it would be good to have that double checked :)