@davidanthoff 2019-02-19
@shashi Apart from the one comment you made, are you on board with this? How do you want the `StringArray` situation to be handled?