@davidanthoff 2019-02-21
@joshday also pinging you on this. Are you the main point person for JuliaDB stuff these days? Would be good to get feedback from the JuliaDB side of things how you want to handle the `StringArray` situation here. Short summary: currently TextParse generally doesn't handle escaped strings properly. This PR here fixes that (I think) for the `Vector{String}` case, but not for `StringArray` columns.

In my mind we have three choices of how to handle this right now: 1) fix the bug only for `Vector{String}`, and leave the old behavior for `StringArray` around, 2) throw a bug if an escaped string comes up with `StringArray`, 3) someone from the JuliaDB side fixes this for `StringArray` as well, and everything is great :) Option 3 would obviously be best, but if there are constraints that make that tricky, it would be good to know whether you prefer 1) or 2) (I'm leaning towards 1), but I think you guys should make that call).