@joshday 2019-02-22
> Are you the main point person for JuliaDB stuff these days?

Yes, but half the time the first thing I do is bother @shashi!

I'd like to look into fixing this for StringArrays (👍 for option 3), and if this PR is good to go I say we merge it (start with option 1).