@davidanthoff 2019-02-22
Sounds good! I think if @shashi officially signs off on this PR we can merge it and then you can fix the `StringArray` case in another PR before we tag a new release.

The thing that made it appear slightly trickier to me for the `StringArray` case is how to do this without an extra temporary allocation. Essentially I think you want to grow the array buffer first (because you know by how much it needs to grow, and then loop over the characters in the source, skipping the ones that are escaped. But `StringArray` right now doesn't seem to provide an API for that kind of thing...