@shashi 2019-02-28
Oh no! Sorry for the hold up here! Been super busy with coursework! :(

This change looks good. I like the separation of concern to `_substring`, but distrust the increase in the number of arguments. But it's OK.

> I generally have changed how the case of `Vector{String}` is handled: in that case this PR is no longer using `StrRange` at all, but is allocating the `String` in the `tryparsenext` method itself. I didn't see a benefit of going through `StrRange`, and it cuts down on the places that need to handle the escaping stuff.

This is a good change, `StrRange` was introduced in Julia 0.6 when we still boxed immutable structs with Strings in them. Ideally we should try to remove it entirely because it's also quite redundant with WeakRefString.

> Or, alternatively, but clearly not ideal, we could also for now just fix the bug for the `Vector{String}` case and then later try to fix it for `StringArray` as well.

This is OK. I couldn't figure out how a user can say "Don't try to un-escape quotes when parsing with the StringArray type", do they just get an error in this case? We should probably not throw here -- otherwise it's really evil towards those who are currently using this to parse strings with escaped quotes!