@davidanthoff 2019-03-17
Some performance numbers here, just want to make sure I record them. I'm running this test on one file that has no escaped quote mark in any string column (this should take the fast path) and one file that has string columns with escaped quotes in it (this will be slower, but returns an incorrect value on `master`).

Here are the numbers for the file without escapes in it:
```
On master:
julia> @time csvread("uniform_string.csv", stringarraytype=Array);
  5.093425 seconds (20.01 M allocations: 1.105 GiB, 43.14% gc time)

With this PR:
julia> @time csvread("uniform_string.csv", stringarraytype=Array);
  5.014085 seconds (60.01 M allocations: 2.297 GiB, 52.21% gc time)
```
So that is all looking good.

The numbers for the file with escapes in it:
```
On master:
julia> @time csvread("uniform_escapedstring.csv", stringarraytype=Array);
  5.744557 seconds (20.01 M allocations: 1.403 GiB, 36.64% gc time)

With this PR:
julia> @time csvread("uniform_escapedstring.csv", stringarraytype=Array);
  9.589759 seconds (80.01 M allocations: 4.383 GiB, 54.45% gc time)
```
So here we have a slow down, but on the other hand, we fix the incorrect result from `master`.

Overall, I think this is ok in terms of performance, in particular because our baseline is sooo much faster than pretty much anything else out there. Plus, the fix pretty much _must_ slow us down.