@davidanthoff 2019-03-17
Alright, assuming tests pass, this would be ready for a merge from my side. There is a fair amount of new stuff in here relative to a few days ago :)

@shashi mentioned somewhere that we should also test the `StringArray` code paths. I agree, but on the other hand, as far as I can tell, those code paths are also not tested on `master` right now, so this PR is not really a regression on that front. So I would say that adding tests for the `StringArray` case could just be a new PR by someone that is independent from this one here.

This now is an implementation of option 1): it fixes the escape handling for the `stringarraytype=Array` case, but not the `stringarraytype=StringArray` case. It doesn't error in the latter case, it just preserves the old behavior for now.

I'm going to wait for one more nod from @shashi that he is ok with a merge here, and then I'm going to merge.