@shashi 2019-03-17
Looking good! Feel free to merge! `stringarraytype=StringArray` case can be handled by probably converting WeakRefString to string and invoking the same code as `Array` case if `escapecount > 0`.