@shashi 2019-03-17
We should look into the performance penalty a bit more before making this the default i.e. in StringArray.

How many string columns with quotes does the csv you're benchmarking have? And how many rows?