@davidanthoff 2019-03-18
We should probably restructure the test code in some way so that all the tests run for all the different source string types we support. Would be cool to also test with say a UTF16 string type, just to make sure that also works.

The numbers from above for the file with escaped strings are a 20 column, 1 million row file, where every string has an escaped quote character in it, so the fast path is never taken. I think that is probably the worst case upper bound, therefore.

There are _a lot_ of things we could do to make this more efficient, the current implementation is quite naive. I'll run my full benchmark suite over it tonight, but I am pretty positive that even with this naive implementation we might still be faster than any other serial CSV reader that I'm testing. But we'll see... In any case, I think there is a fair bit of room for improvement left. But my main objective right now is to just get the correct behavior.