@quinnj 2019-03-18
For StringArray, I actually had an idea the other day to have a `QuotedString` type that would implement the String interface and do the escaping lazily (basically on `show` and `string(x)`). I think performance-wise it could be a big win, because you just do the escaping lazily and avoid the upfront allocations that might not even be needed depending on what the user ends up doing later. I haven't really fleshed the idea out that far, but thought I'd mention it in case someone else wants to explore the idea a little more.